refactor : 관리자 빵집 조회 필터링 추가 및 관리자 빵집 검색 통합, Review에 isNew 추가

+++ src/main/java/com/depromeet/breadmapbackend/domain/bakery/product/report/ProductAddReportImageRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 public interface ProductAddReportImageRepository extends JpaRepository<ProductAddReportImage, Long> {
     boolean existsByBakeryAndIsNewIsTrue(Bakery bakery);
-    long countByBakeryAndIsRegisteredIsTrue(Bakery bakery);
+    Integer countByBakeryAndIsRegisteredIsTrue(Bakery bakery);
     Page<ProductAddReportImage> findPageByBakeryAndIsRegisteredIsTrue(Bakery bakery, Pageable pageable);
     Optional<ProductAddReportImage> findByImageAndProductAddReport(String image, ProductAddReport report);
     Optional<ProductAddReportImage> findByIdAndProductAddReport(Long id, ProductAddReport report);
